Refactor: create SecondaryButton component
