+++ agenta-web/src/components/EvaluationTable/ExactMatchEvaluationTable.tsx
@@ -1,6 +1,6 @@
 import {useState, useEffect} from "react"
 import type {ColumnType} from "antd/es/table"
-import {ExportOutlined, LineChartOutlined} from "@ant-design/icons"
+import {LineChartOutlined} from "@ant-design/icons"
 import {
     Button,
     Card,
@@ -26,14 +26,10 @@ import {useRouter} from "next/router"
 import {EvaluationFlow} from "@/lib/enums"
 import {evaluateWithExactMatch} from "@/lib/services/evaluations"
 import {createUseStyles} from "react-jss"
-import {useAppTheme} from "../Layout/ThemeContextProvider"
 import {exportExactEvaluationData} from "@/lib/helpers/evaluate"
+import SecondaryButton from "../SecondaryButton/SecondaryButton"
 
 const {Title} = Typography
-
-type StyleProps = {
-    themeMode: "dark" | "light"
-}
 
 interface ExactMatchEvaluationTableProps {
     evaluation: any
@@ -100,15 +96,6 @@ const useStyles = createUseStyles({
             color: "#cf1322",
         },
     },
-    exportBtn: ({themeMode}: StyleProps) => ({
-        backgroundColor: themeMode === "dark" ? "#fff" : "#000",
-        color: themeMode === "dark" ? "#000" : "#fff",
-        border: "none",
-        "&:hover": {
-            backgroundColor: themeMode === "dark" ? "rgba(255, 255, 255,0.8)" : "rgba(0, 0, 0,0.8)",
-            color: `${themeMode === "dark" ? "#000" : "#fff"} !important`,
-        },
-    }),
 })
 
 const ExactMatchEvaluationTable: React.FC<ExactMatchEvaluationTableProps> = ({
@@ -116,8 +103,7 @@ const ExactMatchEvaluationTable: React.FC<ExactMatchEvaluationTableProps> = ({
     evaluationScenarios,
     columnsCount,
 }) => {
-    const {appTheme} = useAppTheme()
-    const classes = useStyles({themeMode: appTheme} as StyleProps)
+    const classes = useStyles()
     const router = useRouter()
     const appName = Array.isArray(router.query.app_name)
         ? router.query.app_name[0]
@@ -396,15 +382,12 @@ const ExactMatchEvaluationTable: React.FC<ExactMatchEvaluationTableProps> = ({
                             >
                                 Run Evaluation
                             </Button>
-                            <Button
+                            <SecondaryButton
                                 onClick={() => exportExactEvaluationData(evaluation, rows)}
-                                icon={<ExportOutlined />}
-                                size="large"
-                                className={classes.exportBtn}
                                 disabled={evaluationStatus !== EvaluationFlow.EVALUATION_FINISHED}
                             >
                                 Export results
-                            </Button>
+                            </SecondaryButton>
                         </Space>
                     </Col>
