+++ agenta-web/src/components/EvaluationTable/SimilarityMatchEvaluationTable.tsx
@@ -1,6 +1,6 @@
 import {useState, useEffect} from "react"
 import type {ColumnType} from "antd/es/table"
-import {ExportOutlined, LineChartOutlined} from "@ant-design/icons"
+import {LineChartOutlined} from "@ant-design/icons"
 import {Button, Card, Col, Input, Row, Space, Spin, Statistic, Table, Tag, message} from "antd"
 import {
     updateEvaluationScenario,
@@ -14,14 +14,10 @@ import {EvaluationFlow} from "@/lib/enums"
 import {evaluateWithSimilarityMatch} from "@/lib/services/evaluations"
 import {Typography} from "antd"
 import {createUseStyles} from "react-jss"
-import {useAppTheme} from "../Layout/ThemeContextProvider"
 import {exportSimilarityEvaluationData} from "@/lib/helpers/evaluate"
+import SecondaryButton from "../SecondaryButton/SecondaryButton"
 
 const {Title} = Typography
-
-type StyleProps = {
-    themeMode: "dark" | "light"
-}
 
 interface SimilarityMatchEvaluationTableProps {
     evaluation: any
@@ -92,15 +88,6 @@ const useStyles = createUseStyles({
             color: "#cf1322",
         },
     },
-    exportBtn: ({themeMode}: StyleProps) => ({
-        backgroundColor: themeMode === "dark" ? "#fff" : "#000",
-        color: themeMode === "dark" ? "#000" : "#fff",
-        border: "none",
-        "&:hover": {
-            backgroundColor: themeMode === "dark" ? "rgba(255, 255, 255,0.8)" : "rgba(0, 0, 0,0.8)",
-            color: `${themeMode === "dark" ? "#000" : "#fff"} !important`,
-        },
-    }),
 })
 
 const SimilarityMatchEvaluationTable: React.FC<SimilarityMatchEvaluationTableProps> = ({
@@ -108,8 +95,7 @@ const SimilarityMatchEvaluationTable: React.FC<SimilarityMatchEvaluationTablePro
     evaluationScenarios,
     columnsCount,
 }) => {
-    const {appTheme} = useAppTheme()
-    const classes = useStyles({themeMode: appTheme} as StyleProps)
+    const classes = useStyles()
     const router = useRouter()
     const appName = Array.isArray(router.query.app_name)
         ? router.query.app_name[0]
@@ -443,15 +429,12 @@ const SimilarityMatchEvaluationTable: React.FC<SimilarityMatchEvaluationTablePro
                             >
                                 Run Evaluation
                             </Button>
-                            <Button
+                            <SecondaryButton
                                 onClick={() => exportSimilarityEvaluationData(evaluation, rows)}
-                                icon={<ExportOutlined />}
-                                size="large"
-                                className={classes.exportBtn}
                                 disabled={evaluationStatus !== EvaluationFlow.EVALUATION_FINISHED}
                             >
                                 Export results
-                            </Button>
+                            </SecondaryButton>
                         </Space>
                     </Col>
